+++ contracts/contracts/strategies/crosschain/CrossChainRemoteStrategy.sol
@@ -71,8 +71,8 @@ contract CrossChainRemoteStrategy is
     function initialize(
         address _strategist,
         address _operator,
-        uint32 _minFinalityThreshold,
-        uint32 _feePremiumBps
+        uint16 _minFinalityThreshold,
+        uint16 _feePremiumBps
     ) external virtual onlyGovernor initializer {
         _initialize(_operator, _minFinalityThreshold, _feePremiumBps);
         _setStrategistAddr(_strategist);
